Fix atomic function on win32
This is a nasty hack; basically the function iris_atomics_fetch_and_add()
was implemented using InterlockedAdd(); this is Vista+ only. The only user
of the function only added 1, so I changed the function to _fetch_and_inc().
The OSX version I believe works as add_and_fetch() rather than
fetch_and_add() so I used the same hack as the Windows version - subtract 1
again from the result.

Nasty and nastier. But it seems to work.