Fix scheduler foreach bug
This was causing an infinite loop in iris_receiver_destroy() in certain
circumstances. In iris_scheduler_foreach_rrobin_cb() the
'continue_flag' was initially FALSE; this meant if a thread's queue
was empty, the flag wouldn't be touched and the foreach would therefore
be aborted (due to returning FALSE). Of course if a message was hiding
out in a later queue, the receiver would hang waiting for it to be
unqueued.